Fix export of DENY OUT rules
Use a modified version of UFWCommandRule.get_command() to correctly
generate the command string for a DENY OUT rule. The new method
basically includes the direction of the rule in the command string.

As an example, for the following rule:

Action | Direction | Proto | Source | Port/App | Destination | Port/App
-----------------------------------------------------------------------
DENY   | OUT       | *     | *      | *        | 1.1.1.1     | *

The original version of the method outputs:
'deny to 1.1.1.1'

while the modified version outputs:
'deny out to 1.1.1.1'