Fix crash when no log in supported location can be found
This can happen for example on systems that use Journal
from systemd.

In this case, ufw-gtk exits with a traceback containing:
IOError: [Errno 2] No such file or directory: '/var/log/messages.log'
(this is the last log file tried).

The patch works around the issue by handling the error
and disabling the widget in the "Events" tab.